+++ src/Bridge/Sylius/config/services.php
@@ -8,6 +8,7 @@ use JoliCode\MediaBundle\Bridge\Sylius\Admin\Form\Type\MediaChoiceType;
 use JoliCode\MediaBundle\Bridge\Sylius\Admin\Form\Type\UploadType;
 use JoliCode\MediaBundle\Bridge\Sylius\Admin\Grid\MediaGrid;
 use JoliCode\MediaBundle\Bridge\Sylius\Admin\Grid\Provider\MediaGridProvider;
+use JoliCode\MediaBundle\Bridge\Sylius\Asset\Package;
 use JoliCode\MediaBundle\Bridge\Sylius\Config\Config;
 use JoliCode\MediaBundle\Bridge\Sylius\Uploader\ImageUploader;
 use JoliCode\MediaBundle\Bridge\Twig\JoliMediaAdminExtension;
@@ -15,6 +16,12 @@ use Sylius\Component\Core\Uploader\ImageUploaderInterface;
 
 return static function (ContainerConfigurator $container): void {
     $services = $container->services();
+
+    $services
+        ->set('joli_media_sylius_admin.assets_package', Package::class)
+        ->args([service('request_stack')])
+        ->tag('assets.package', ['package' => Package::NAME])
+    ;
 
     $services
         ->set('joli_media_sylius_admin.config', Config::class)