+++ src/Bridge/Sylius/src/Asset/Package.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace JoliCode\MediaBundle\Bridge\Sylius\Asset;
+
+use Symfony\Component\Asset\Context\RequestStackContext;
+use Symfony\Component\Asset\PackageInterface;
+use Symfony\Component\Asset\PathPackage;
+use Symfony\Component\Asset\VersionStrategy\JsonManifestVersionStrategy;
+use Symfony\Component\HttpFoundation\RequestStack;
+
+final readonly class Package implements PackageInterface
+{
+    public const NAME = 'joli-media-sylius-admin';
+
+    private PackageInterface $package;
+
+    public function __construct(RequestStack $requestStack)
+    {
+        $this->package = new PathPackage(
+            '/bundles/jolimediasylius',
+            new JsonManifestVersionStrategy(__DIR__ . '/../../public/manifest.json'),
+            new RequestStackContext($requestStack)
+        );
+    }
+
+    public function getUrl(string $assetPath): string
+    {
+        return $this->package->getUrl($assetPath);
+    }
+
+    public function getVersion(string $assetPath): string
+    {
+        return $this->package->getVersion($assetPath);
+    }
+}